refactor: set grpc client to common use
