+++ pkg/grpc.go
@@ -1,4 +1,4 @@
-package xpla
+package pkg
 
 import (
 	"context"
