+++ pkg/xpla/mock/mocks.go
@@ -2,7 +2,7 @@ package mock
 
 import (
 	ibc_types "github.com/cosmos/ibc-go/v3/modules/apps/transfer/types"
-	"github.com/dezswap/dezswap-api/pkg/xpla"
+	"github.com/dezswap/dezswap-api/pkg"
 	"github.com/stretchr/testify/mock"
 )
 
@@ -10,25 +10,25 @@ type GrpcClientMock struct {
 	*mock.Mock
 }
 
-var _ xpla.GrpcClient = &GrpcClientMock{}
+var _ pkg.GrpcClient = &GrpcClientMock{}
 
 func NewGrpcClientMock() *GrpcClientMock {
 	return &GrpcClientMock{&mock.Mock{}}
 }
 
-// QueryContract implements xpla.GrpcClient
+// QueryContract implements pkg.GrpcClient
 func (g *GrpcClientMock) QueryContract(addr string, query []byte, height uint64) ([]byte, error) {
 	args := g.MethodCalled("QueryContract", addr, query, height)
 	return args.Get(0).([]byte), args.Error(1)
 }
 
-// SyncedHeight implements xpla.GrpcClient
+// SyncedHeight implements pkg.GrpcClient
 func (g *GrpcClientMock) SyncedHeight() (uint64, error) {
 	args := g.MethodCalled("SyncedHeight")
 	return args.Get(0).(uint64), args.Error(1)
 }
 
-// QueryIbcDenomTrace implements xpla.GrpcClient
+// QueryIbcDenomTrace implements pkg.GrpcClient
 func (g *GrpcClientMock) QueryIbcDenomTrace(hash string) (*ibc_types.DenomTrace, error) {
 	args := g.MethodCalled("QueryIbcDenomTrace")
 	return args.Get(0).(*ibc_types.DenomTrace), args.Error(1)
